+++ dictionaries/dictionary.d.ts
@@ -1,0 +1,4 @@
+declare module '*/dictionary.js' {
+  const content: string;
+  export default content;
+}